+++ packages/terra-popup/src/terra-dev-site/doc/example/PopupAttachmentBehavior.jsx
@@ -1,35 +1,62 @@
-import React from 'react';
-import PopupCommonExample from 'terra-popup/lib/terra-dev-site/doc/common/PopupCommonExample';
+import React, { useRef, useState } from 'react';
+import Popup from 'terra-popup';
+import Button from 'terra-button';
+import Placeholder from 'terra-doc-template/lib/Placeholder';
 import classNames from 'classnames/bind';
 import styles from './PopupAttachmentBehavior.module.scss';
 
 const cx = classNames.bind(styles);
 
-class PopupAttachmentBehavior extends React.Component {
-  constructor(props) {
-    super(props);
+function PopupAttachmentBehavior() {
+  const buttonElement = useRef();
+  const [open, setOpen] = useState(false);
+  const [contentBehavior, setContentBehavior] = useState('auto');
 
-    this.state = { behavior: 'auto' };
-    this.handleChange = this.handleChange.bind(this);
-  }
+  const setButtonNode = (node) => {
+    buttonElement.current = node;
+  };
 
-  handleChange(event) {
-    this.setState({ behavior: event.target.value });
-  }
+  const getButtonNode = () => buttonElement.current;
 
-  render() {
-    return (
+  const handleButtonClick = () => {
+    setOpen(true);
+  };
+
+  const handleRequestClose = () => {
+    setOpen(false);
+  };
+
+  const handleChange = (event) => {
+    setContentBehavior(event.target.value);
+  };
+
+  return (
+    <React.Fragment>
       <div>
         <label htmlFor="popup-behavior">Attachment Behavior</label>
-        <select id="popup-behavior" onChange={this.handleChange} value={this.state.behavior} className={cx('selection-list')}>
+        <select id="popup-behavior" onChange={handleChange} value={contentBehavior} className={cx('selection-list')}>
           <option value="auto">Auto</option>
           <option value="flip">Flip</option>
           <option value="push">Push</option>
         </select>
-        <PopupCommonExample title="Open Popup" attachmentBehavior={this.state.behavior} />
+        <Button
+          className={cx('popup-wrapper')}
+          text="Open Popup"
+          onClick={handleButtonClick}
+          refCallback={setButtonNode}
+        />
       </div>
-    );
-  }
+      <Popup
+        attachmentBehavior={contentBehavior}
+        contentAttachment="top center"
+        isOpen={open}
+        targetRef={getButtonNode}
+        onRequestClose={handleRequestClose}
+      >
+        <Placeholder title="Popup Content" />
+      </Popup>
+    </React.Fragment>
+  );
 }
 
 export default PopupAttachmentBehavior;