+++ packages/terra-aggregator/CHANGELOG.md
@@ -6,6 +6,8 @@ Unreleased
 ### Changed
 * Added `aria-label` in examples.
 
+4.35.0 - (December 10, 2019)
+------------------
 ### Added
 * Added reference wdio screenshots for "orion-fusion-theme".